Refactored connection dialogue box; no more bug when submitting form/ports; creates new Row for new connection
