+++ src/components/ConnectionDialogBox.jsx
@@ -18,33 +18,20 @@ import { styled } from '@mui/system';
 import AddIcon from '@mui/icons-material/Add';
 
 
-// TO DO: confirm apiUrl for production
-const apiUrl =
-  process.env.NODE_ENV === 'production'
-    ? 'https://api.kmon.com'
-    : 'http://localhost:3010';
-
-
-const ConnectionDialogBox = () => {
+const ConnectionDialogBox = (props) => {
 
   /************** Component States ****************/
 
   // states for port entry
-  const [portIsClicked, setPortIsClicked] = useState(false);
   const [portIsValid, setPortIsValid] = useState(true)
   const [helperText, setHelperText] = useState(null)
   // clear hard-coded faults for production
-  // const [ports, setPorts] = useState(['1234', '2345', '3456']);
   // form states
-  const [submitting, setSubmitting] = useState(false);
+  const [formData, setFormData] = props.formData;
+  const [portIsClicked, setPortIsClicked] = props.portIsClicked;
+  const [submitting, setSubmitting] = props.submitting;
+
   // clear hard-coded formData for production;
-  const [formData, setFormData] = useState({
-    clusterName: 'Demo',
-    serverURI: '10.0.10.137',
-    ports: ['8994', '8995', '8996', '8997', '8998', '8999'],
-    apiKey: '',
-    apiSecret: '',
-  });
   const [open, setOpen] = useState(false);
   // alert props to display in case of invalid form Input
   const [alertProps, setAlertProps] = useState({
@@ -52,7 +39,6 @@ const ConnectionDialogBox = () => {
     height: 0,
     message: ''
   })
-
   /************** Event Handlers *************/
 
   // Handle change for form input
@@ -67,63 +53,8 @@ const ConnectionDialogBox = () => {
   // TO DO: 
   // - update api call
   // - need to fix up ports input event handlers. Currently, clicking "enter" with a valid port updates port numbers in state but also serves sets the state for "portIsValid" to false, which changes the appearance of the input box.
-  const handleSubmit = async (event) => {
-    console.log(portIsClicked);
-    event.preventDefault();
-    if (!portIsClicked) {
-      // check if form is valid, otherwise display alert.
-      if (!formData.clusterName || !formData.serverURI || !formData.ports.length) {
-        setAlertProps({
-          visibility: 'visible',
-          marginTop: '15px',
-          height: '100%',
-          message: 'Please provide cluster name, server URI, and port numbers'
-        })
-      }
-      else
-        try {
-          const response = await fetch(`${apiUrl}/api/createConnection`, {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({
-              ...formData
-            }),
-          });
-          console.log(response);
 
-          if (response.ok) {
-            const data = await response.json();
-            setSubmitting(false);
-            setFormData({
-              clusterName: '',
-              serverURI: '',
-              ports: [],
-              apiKey: '',
-              apiSecret: '',
-            });
-            console.log('data submitted! data: ', data)
-          } else {
-            console.log('Failed to save credentials');
-          }
-        } catch (error) {
-          console.log('Error in Credential Form: ', error);
-        } finally {
-          setSubmitting(false);
-          setFormData({
-            clusterName: '',
-            serverURI: '',
-            ports: [],
-            apiKey: '',
-            apiSecret: '',
-          });
-        }
-    }
-    else {
-      event.preventDefault();
-    }
-  };
+  const handleSubmit = props.handleSubmit;
 
   const handleSubmitKey = (event) => {
